Actually update velocities

+++ shared/include/EigenTypes.h
@@ -43,37 +43,37 @@ namespace Eigen {
     using SparseMatrixd = Eigen::SparseMatrix<double>;
 
 	// Forgive us for this we had no other choice lmao
-	using TensorXV = std::array<Eigen::Matrix<float, TENSOR_XV_Y, TENSOR_XV_Z>, TENSOR_XV_X>;
-	using TensorYV = std::array<Eigen::Matrix<float, TENSOR_YV_Y, TENSOR_YV_Z>, TENSOR_YV_X>;
-	using TensorZV = std::array<Eigen::Matrix<float, TENSOR_ZV_Y, TENSOR_ZV_Z>, TENSOR_ZV_X>;
-	using TensorP = std::array<Eigen::Matrix<float, TENSOR_P_Y, TENSOR_P_Z>, TENSOR_P_X>;
+	using TensorXV = std::array<Eigen::Matrix<double, TENSOR_XV_Y, TENSOR_XV_Z>, TENSOR_XV_X>;
+	using TensorYV = std::array<Eigen::Matrix<double, TENSOR_YV_Y, TENSOR_YV_Z>, TENSOR_YV_X>;
+	using TensorZV = std::array<Eigen::Matrix<double, TENSOR_ZV_Y, TENSOR_ZV_Z>, TENSOR_ZV_X>;
+	using TensorP = std::array<Eigen::Matrix<double, TENSOR_P_Y, TENSOR_P_Z>, TENSOR_P_X>;
 	using TensorPB = std::array<std::array<std::array<Eigen::Matrix<double, 6, 6>, TENSOR_P_Z>, TENSOR_P_Y>, TENSOR_P_X>;
 
 }
 
 template <int X, int Y, int Z>
-inline bool validCoordinates(std::array<Eigen::Matrix<float, Y, Z>, X>& tensor, int x, int y, int z) {
+inline bool validCoordinates(std::array<Eigen::Matrix<double, Y, Z>, X>& tensor, int x, int y, int z) {
 	return ((0 <= x && x < X) && (0 <= y && y < Y) && (0 <= z && z < Z));
 }
 
 template <int X, int Y, int Z>
-inline bool validCoordinates(std::array<Eigen::Matrix<float, Y, Z>, X>& tensor, Eigen::Vector3i idx) {
+inline bool validCoordinates(std::array<Eigen::Matrix<double, Y, Z>, X>& tensor, Eigen::Vector3i idx) {
 	return validCoordinates(tensor, idx(0), idx(1), idx(2));
 }
 
 template <int X, int Y, int Z>
-inline float tensorAtOrZero(std::array<Eigen::Matrix<float, Y, Z>, X>& tensor, int x, int y, int z) {
+inline double tensorAtOrZero(std::array<Eigen::Matrix<double, Y, Z>, X>& tensor, int x, int y, int z) {
 	if (validCoordinates(tensor, x, y, z)) return tensorAt(tensor, x, y, z);
 	return 0.0;
 }
 
 template <int X, int Y, int Z>
-inline float& tensorAt(std::array<Eigen::Matrix<float, Y, Z>, X>& tensor, Eigen::Vector3i idx) {
+inline double& tensorAt(std::array<Eigen::Matrix<double, Y, Z>, X>& tensor, Eigen::Vector3i idx) {
 	return tensorAt(tensor, idx(0), idx(1), idx(2));
 }
 
 template <int X, int Y, int Z>
-inline float& tensorAt(std::array<Eigen::Matrix<float, Y, Z>, X>& tensor, int x, int y, int z) {
+inline double& tensorAt(std::array<Eigen::Matrix<double, Y, Z>, X>& tensor, int x, int y, int z) {
 	assert(0 <= x && x < X);
 	assert(0 <= y && y < Y);
 	assert(0 <= z && z < Z);
